progressbar/iocopy: initial implementations
This commit adds the progressbar and iocopy files. progressbar exposes
an API for multiline progress bars, which is much simpler than the
current ioprogress API. Ideally the current API can be phased out in
favor of this one in the future.

iocopy provides a utility for performing parallel io.Copy calls, and
using the code in progressbar to draw progress bars about the transfers.